% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_word.R
\name{check_word}
\alias{check_word}
\title{Check whether the word is valid}
\usage{
check_word(word)
}
\arguments{
\item{word}{}
}
\value{
T or F
}
\description{
Check whether the word is valid
}
\examples{
check_word("love")
}
