% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{Dictionary}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
online
}
\usage{
data(dictionary)
}
\description{
Stored all the words in a dictionary
}
\examples{
data(dictionary)
}
\references{
Phyllis Li
}
\keyword{datasets}
