% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_word.R
\name{find_corrections}
\alias{find_corrections}
\title{Find the length of the word}
\usage{
find_corrections(wrongWord)
}
\arguments{
\item{word}{}
}
\value{
the length of the word
}
\description{
Find the length of the word
}
\examples{
find_corrections("lovve")
}
