% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_word.R
\name{find_length}
\alias{find_length}
\title{Find the length of the word}
\usage{
find_length(word)
}
\arguments{
\item{word}{}
}
\value{
the length of the word
}
\description{
Find the length of the word
}
\examples{
find_length("love")
}
